requirements.txt
Flask==2.3.2
pdfplumber==0.9.0
arabic-reshaper==3.0.0
python-bidi==0.4.2
XlsxWriter==3.1.9